Rename to RightAccessoryButton
