+++ Tests/ReferenceImages_64/StripePaymentSheetTests.RightAccessoryButtonSnapshotTests/testRightAccessoryButtonSnapshotTests_editAndReturnError_@3x.png
