+++ Tests/ReferenceImages_64/StripePaymentSheetTests.RightAccessoryButtonSnapshotTests/testRightAccessoryButtonSnapshotTests_edit_customAppearanceAndReturnError_@3x.png
