+++ Tests/ReferenceImages_64/StripePaymentSheetTests.RightAccessoryButtonSnapshotTests/testRightAccessoryButtonSnapshotTests_viewMoreAndReturnError_@3x.png
