+++ Tests/ReferenceImages_64/StripePaymentSheetTests.RightAccessoryButtonSnapshotTests/testRightAccessoryButtonSnapshotTests_viewMore_customAppearanceAndReturnError_@3x.png
